+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT UP
 
 - Form of the beast effect
+- Tweaks to Thunderclap and Sword Burst spells if using midi effects to better use midi features.
 
 # 3.4.17